XSLT in css-core can be moved to css-utils
because it is not used in any other css submodules
