--- pipeline-braille-utils/css-utils/css-core/src/test/java/CSSCoreTest.java
@@ -1,93 +1,0 @@
-import java.io.File;
-import java.util.Hashtable;
-
-import javax.inject.Inject;
-
-import com.google.common.base.Optional;
-
-import org.daisy.maven.xspec.TestResults;
-import org.daisy.maven.xspec.XSpecRunner;
-
-import org.daisy.pipeline.braille.common.TextTransform;
-import org.daisy.pipeline.braille.common.TextTransform.ContextUnawareTextTransform;
-
-import static org.daisy.pipeline.pax.exam.Options.brailleModule;
-import static org.daisy.pipeline.pax.exam.Options.bundlesAndDependencies;
-import static org.daisy.pipeline.pax.exam.Options.domTraversalPackage;
-import static org.daisy.pipeline.pax.exam.Options.felixDeclarativeServices;
-import static org.daisy.pipeline.pax.exam.Options.logbackBundles;
-import static org.daisy.pipeline.pax.exam.Options.logbackConfigFile;
-import static org.daisy.pipeline.pax.exam.Options.thisBundle;
-import static org.daisy.pipeline.pax.exam.Options.xspecBundles;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.assertEquals;
-
-import org.ops4j.pax.exam.Configuration;
-import org.ops4j.pax.exam.junit.PaxExam;
-import org.ops4j.pax.exam.Option;
-import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
-import org.ops4j.pax.exam.spi.reactors.PerClass;
-import org.ops4j.pax.exam.util.PathUtils;
-
-import static org.ops4j.pax.exam.CoreOptions.junitBundles;
-import static org.ops4j.pax.exam.CoreOptions.mavenBundle;
-import static org.ops4j.pax.exam.CoreOptions.options;
-
-import org.osgi.framework.BundleContext;
-
-@RunWith(PaxExam.class)
-@ExamReactorStrategy(PerClass.class)
-public class CSSCoreTest {
-	
-	@Configuration
-	public Option[] config() {
-		return options(
-			logbackConfigFile(),
-			domTraversalPackage(),
-			logbackBundles(),
-			felixDeclarativeServices(),
-			mavenBundle().groupId("org.apache.servicemix.bundles").artifactId("org.apache.servicemix.bundles.antlr-runtime").versionAsInProject(),
-			mavenBundle().groupId("org.daisy.libs").artifactId("jstyleparser").versionAsInProject(),
-			bundlesAndDependencies("org.daisy.pipeline.calabash-adapter"),
-			brailleModule("common-utils"),
-			thisBundle("org.daisy.pipeline.modules.braille", "css-core"),
-			xspecBundles(),
-			junitBundles()
-		);
-	}
-	
-	@Inject
-	private BundleContext context;
-	
-	@Before
-	public void registerUppercaseTranslatorProvider() {
-		context.registerService(
-			TextTransform.Provider.class.getName(),
-			new TextTransform.Provider<TextTransform>() {
-				public Iterable<TextTransform> get(String query) {
-					return Optional.<TextTransform>fromNullable(
-						query.equals("(uppercase)")
-							? new ContextUnawareTextTransform() {
-								public String transform(String text) {
-									return text.toUpperCase(); }}
-							: null).asSet(); }},
-			new Hashtable<String,Object>());
-	}
-	
-	@Inject
-	private XSpecRunner xspecRunner;
-	
-	@Test
-	public void runXSpec() throws Exception {
-		File baseDir = new File(PathUtils.getBaseDir());
-		File testsDir = new File(baseDir, "src/test/xspec");
-		File reportsDir = new File(baseDir, "target/surefire-reports");
-		reportsDir.mkdirs();
-		TestResults result = xspecRunner.run(testsDir, reportsDir);
-		assertEquals("Number of failures and errors should be zero", 0L, result.getFailures() + result.getErrors());
-	}
-}
