--- pipeline-braille-utils/css-utils/css-core/src/test/resources/exam.properties
@@ -1,1 +1,0 @@
-pax.exam.logging = none
