--- pipeline-braille-utils/css-utils/css-core/src/test/resources/logback.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-  
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-    </encoder>
-  </appender>
-   
-  <appender name="TEST_LOG" class="ch.qos.logback.core.FileAppender">
-    <file>target/test.log</file>
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-    </encoder>
-  </appender>
-  
-  <root level="WARN">
-    <appender-ref ref="TEST_LOG"/>
-  </root>
-  
-  <logger name="org.daisy" level="DEBUG"/>
-  
-</configuration>
