+++ pipeline-braille-utils/css-utils/css-utils/src/main/resources/xml/library.xsl
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" version="2.0">
     
-    <!--
-        css-core [2.0.0,3.0.0)
-    -->
-    <xsl:import href="http://www.daisy.org/pipeline/modules/braille/css-core/braille-css.xsl"/>
+    <xsl:import href="braille-css.xsl"/>
     
 </xsl:stylesheet>
